+++ __tests__/unit/plots/pie/interaction-spec.ts
@@ -4,7 +4,7 @@ import { deepMix } from '@antv/util';
 import { delay } from '../../../utils/delay';
 import { createDiv } from '../../../utils/dom';
 import { Pie } from '../../../../src';
-import { StatisticAction } from '../../../../src/plots/pie/interaction/pie-statistic-action';
+import { StatisticAction } from '../../../../src/plots/pie/interactions/pie-statistic-action';
 
 describe('register interaction', () => {
   it('创建 "pie-statistic" action', () => {
