+++ src/plots/pie/interactions/pie-statistic-action.ts
